+++ Functions/Public/Get-LockedOutADUsers.ps1
@@ -13,10 +13,18 @@
         - LockoutTime
         - [CONDITIONAL] LockoutSource (See Parameter -IncludeLockoutSource)
 .PARAMETER IncludeLockoutSource
-    This parameter is optional. If specified, the LockoutSource will be returned for each user.
-    This parameter requires you have rights to query your domain controller for these logs.
+    This parameter is optional. If specified, the LockoutSource will be returned for each user. This parameter requires you have rights to query your domain controller for these logs.
     Specifying this parameter will first make a test query to your domain controller to confirm you have rights to query these logs.
     If you do not have rights to query these logs, this parameter will be ignored.
+
+    So far, the SIMPLEST way to grant a domain user rights to read these logs from the correct Domain Controller (that I know of) involves the following:
+        - Add the user (or group) to the 'Event Log Readers' group on the domain controller.
+        - Grant the user (or group) READ rights to the registry key HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Services\Eventlog\Security on the domain controller.
+    Both of the above are required to have rights to the DC's Security logs. Membership to the 'Event Log Readers' group only will NOT grant read rights to the Security logs.
+    See: https://social.technet.microsoft.com/Forums/lync/en-US/b72162d1-2c86-4d1a-9727-ec7269814cc4/getwinevent-with-nonadministrative-user?forum=winserverpowershell
+
+    This will grant the user (or group) rights to read the Event Logs on the domain controller, including the Security logs. This is the only registry key that is required to read the Security Event Logs.
+    If you do not have these rights, you will not be able to read the Lockout (4740) logs.
 .INPUTS
     System.String
         This function does not accept pipeline data. The values for all parameters must be specified.
@@ -60,9 +68,10 @@
 Function Get-LockedOutADUsers() {
     [CmdletBinding()]
     param(
-        [System.String[]] $Properties,
+        [Switch] $IncludeLockoutSource,
+        
+        [System.String[]] $Properties
 
-        [Switch] $IncludeLockoutSource
     )
 
     Begin {